requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0

# Vector database and embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Data processing
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Caching (optional Redis)
redis>=5.0.0

# Monitoring and logging
prometheus-client>=0.19.0

# Utilities
tqdm>=4.66.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
